Add files via upload
Set of functions from both HW1 and HW2. This commit also contains the myrobot class. Modeling a robot as an object makes sense as we can define certain properties such as dimensions, number of joints, joint positions, screw axes and attribute to the object. Furthermore, we can track the state of robot by accessing and updating these attributes. FK_space and J_space claculate forward kinmeatics and Jacobian, resepctively but are agnostic of a particular robot. They operate directly with a general member of the myrobot class which is passed as an input parameter.